[BUGFIX] Update select filter popovers for Bootstrap 5
Popovers are attached to the body by default so we change this to the
parent element of the toggle button.

Also we look up the popover in all siblings since now the hidden form
field is the element next to the button.

Finally the class of the popover content has changed.